# src/AppBundle/Entity/Job.php
<?php

namespace AppBundle\Entity;

use Doctrine\ORM\Mapping as ORM;

/**
 * @ORM\Entity(repositoryClass="AppBundle\Repository\Job")
 * @ORM\Table(name="jobs")
 */
class Job
{
    /**
     * @ORM\GeneratedValue(strategy="IDENTITY")
     * @ORM\Column(type="integer", length=11)
     * @ORM\Id
     * @var int
     */
    protected $id;

    /**
     * @ORM\Column(type="integer", length=11)
     * @var int
     */
    protected $user_id;

    /**
     * @ORM\Column(type="string", length=64)
     * @var string
     */
    protected $title;

    /**
     * @ORM\Column(type="text")
     * @var string
     */
    protected $description;

    /**
     * @ORM\Column(type="integer")
     * @var int
     */
    protected $salary;

    /**
     * @ORM\Column(type="string")
     * @var string
     */
    protected $postcode;

    /**
     * @ORM\Column(type="datetime")
     * @var \DateTime
     */
    protected $created;

    public function setId(int $id)
    {
        $this->id = $id;

        return $this;
    }

    public function setUserId(int $userId)
    {
        $this->user_id = $userId;

        return $this;
    }

    public function setTitle(string $title)
    {
        $this->title = $title;

        return $this;
    }

    public function setDescription(string $description)
    {
        $this->description = $description;

        return $this;
    }

    public function setSalary(int $salary)
    {
        $this->salary = $salary;

        return $this;
    }

    public function setPostcode(string $postcode)
    {
        $this->postcode = $postcode;

        return $this;
    }

    public function setCreated(\DateTime $created)
    {
        $this->created = $created;

        return $this;
    }
}