# web/swagger/spec.json
{
    "swagger": "2.0",
    "info": {
        "title": "HackerGames API",
        "version": "1.0.0"
    },
    "paths": {
        "/jobs-search": {
            "post": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get list of jobs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "JobSearch",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/JobSearch"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of rows to fetch",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Number of rows to skip",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List returned"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/jobs": {
            "post": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Post a new job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Job",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Job"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "New job posted"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Job": {
            "properties": {
                "user_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "salary": {
                    "type": "integer"
                },
                "postcode": {
                    "type": "string"
                }
            }
        },
        "JobSearch": {
            "properties": {
                "skillIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    }
}